SQL_Commands_Error_Checking_Mini_Interpreter
SQL COMMANDS ERROR CHECKING (MINI-INTERPRETER)
As the name of this project suggests it basically reads SQL commands from a text files and checks for any syntax errors and if an error is found it is shown on the screen. I have done only three major SQL commands namely CREATE TABLE, ALTER TABLE, and INSERT INTO. This program shows what an interpreter actually does in the background.